sorry, mistaken check in....

+++ libmp3lame/machine.h
@@ -88,7 +88,7 @@
 #	define WIN32_LEAN_AND_MEAN
 #	include <windows.h>
 #else
-	typedef float FLOAT;
+	typedef double FLOAT;
 #endif
 
 /*
